functional example_generic_event_callback
